+++ Listable/Sources/Layout/List/DefaultListLayout.swift
@@ -449,7 +449,14 @@ final class DefaultListLayout : ListLayout
             // Add additional padding from config.
             
             if isLast == false {
-                let additionalSectionSpacing = hasSectionFooter ? layout.interSectionSpacingWithFooter : layout.interSectionSpacingWithNoFooter
+                let additionalSectionSpacing: CGFloat
+                if let customInterSectionSpacing = section.layout.customInterSectionSpacing {
+                    additionalSectionSpacing = customInterSectionSpacing
+                } else {
+                    additionalSectionSpacing = hasSectionFooter
+                        ? layout.interSectionSpacingWithFooter
+                        : layout.interSectionSpacingWithNoFooter
+                }
                 
                 lastSectionMaxY += additionalSectionSpacing
                 lastContentMaxY += additionalSectionSpacing